+++ main.py
@@ -4,7 +4,7 @@ import pathlib
 from threading import Lock
 from typing import Generator
 
-from kytos.core import KytosEvent, KytosNApp, log, rest
+from kytos.core import KytosNApp, log, rest
 from kytos.core.helpers import listen_to, load_spec, validate_openapi
 from kytos.core.rest_api import (HTTPException, JSONResponse, Request,
                                  get_json_or_400)
@@ -25,8 +25,6 @@ class Main(KytosNApp):
         self.graph = KytosGraph()
         self._topology = None
         self._lock = Lock()
-        self._topology_updated_at = None
-        self._links_updated_at = {}
 
     def execute(self):
         """Do nothing."""
@@ -141,8 +139,8 @@ class Main(KytosNApp):
 
     @listen_to(
         "kytos.topology.updated",
-        "kytos/topology.current",
         "kytos/topology.topology_loaded",
+        pool="dynamic_single"
     )
     def on_topology_updated(self, event):
         """Update the graph when the network topology is updated."""
@@ -154,40 +152,8 @@ class Main(KytosNApp):
             return
         topology = event.content["topology"]
         with self._lock:
-            if (
-                self._topology_updated_at
-                and self._topology_updated_at > event.timestamp
-            ):
-                return
             self._topology = topology
-            self._topology_updated_at = event.timestamp
             self.graph.update_topology(topology)
         switches = list(topology.switches.keys())
         links = list(topology.links.keys())
         log.debug(f"Topology graph updated with switches: {switches}, links: {links}.")
-
-    def update_links_metadata_changed(self, event) -> None:
-        """Update the graph when links' metadata are added or removed."""
-        link = event.content["link"]
-        try:
-            with self._lock:
-                if (
-                    link.id in self._links_updated_at
-                    and self._links_updated_at[link.id] > event.timestamp
-                ):
-                    return
-                self.graph.update_link_metadata(link)
-                self._links_updated_at[link.id] = event.timestamp
-            metadata = event.content["metadata"]
-            log.debug(f"Topology graph updated link id: {link.id} metadata: {metadata}")
-        except KeyError as exc:
-            log.warning(
-                f"Unexpected KeyError {str(exc)} on event {event}."
-                " pathfinder will reconciliate the topology"
-            )
-            self.controller.buffers.app.put(KytosEvent(name="kytos/topology.get"))
-
-    @listen_to("kytos/topology.links.metadata.(added|removed)")
-    def on_links_metadata_changed(self, event):
-        """Update the graph when links' metadata are added or removed."""
-        self.update_links_metadata_changed(event)